+++ isaaclab_arena/environments/arena_env_graph_spec.py
@@ -11,6 +11,7 @@ from typing import TYPE_CHECKING, Any
 
 from pydantic import BaseModel, Field, SerializeAsAny, ValidationInfo, field_validator, model_validator
 
+from isaaclab_arena.assets.registries import TaskRegistry
 from isaaclab_arena.environments.arena_env_graph_types import (
     ArenaEnvGraphCliOverrideSpec,
     ArenaEnvGraphNodeSpec,
@@ -29,7 +30,9 @@ from isaaclab_arena.environments.graph_spec_utils import (
     assert_spatial_constraint_shapes,
     assert_task_wiring,
     assert_unique_ids,
+    spatial_constraint_is_spawn_pose,
 )
+from isaaclab_arena.tasks.task_transition import Relocate, TaskTransition
 
 if TYPE_CHECKING:
     import argparse
@@ -154,16 +157,61 @@ class ArenaEnvGraphSpec(BaseModel):
             yaml.safe_dump(self.to_dict(), f, sort_keys=False)
 
     def resolve_constraints(self, env_name: str | None = None) -> ArenaEnvGraphSpec:
-        """Chain this unresolved graph's tasks into the full state graph and wire each task.
+        """Chain this unresolved graph's task success conditions into its full state specs.
 
-        Topology is implicit in the sequential task list (task ``i`` carries ``state_spec_i`` to
-        ``state_spec_{i+1}``), so resolving only fills the per-state constraints: each task's success
-        condition is applied to the previous state to derive the next one. Returns a new,
-        fully-wired ``ArenaEnvGraphSpec`` (strict-validated). ``env_name`` defaults to this spec's.
+        The upstream emits a partially-wired graph: nodes, the ordered task list, and only the initial
+        state ``state_spec_0`` -- tasks are not yet wired to states. The topology is implicit in the
+        sequential task list (task ``i`` carries ``state_spec_i`` to ``state_spec_{i+1}``), so resolving
+        only fills the per-state constraints: each task's success condition is applied to the previous
+        state to derive the next one. ``N`` tasks therefore yield ``N+1`` state specs. Returns a new,
+        fully-wired, strict-validated ``ArenaEnvGraphSpec``; ``env_name`` defaults to this spec's.
         """
-        from isaaclab_arena.environments.state_spec_resolver import resolve_constraints
+        assert len(self.state_specs) == 1, (
+            f"unresolved graph must define exactly the initial state (state_spec_0); got {len(self.state_specs)} state"
+            " specs"
+        )
+        embodiment_id = _get_embodiment_id_from_nodes(self.nodes)
 
-        return resolve_constraints(self, env_name=env_name)
+        transitions = [_get_task_state_transition(task) for task in self.tasks]
+
+        # state_spec_0 is the given initial state; chain the rest off the task list.
+        states: list[ArenaEnvGraphStateSpec] = [self.state_specs[0]]
+        out_tasks: list[ArenaEnvGraphTaskSpec] = []
+        num_tasks = len(self.tasks)
+        for i, task in enumerate(self.tasks):
+            new_state_id = f"state_spec_{i + 1}"
+            is_final_state = i == num_tasks - 1
+            # A success state is both a postcondition of the task that just ran and a precondition of the
+            # next one, so it asserts reachability of both: the completed task's target (e.g. a place
+            # destination), and -- when a next task exists -- that task's subject (the next thing to act on).
+            reach_targets_postcondition = [transitions[i].reach_target_on_success]
+            reach_targets_precondition = []
+            # final state has no precondition.
+            if not is_final_state:
+                reach_targets_precondition = [transitions[i + 1].subject]
+            reach_targets = reach_targets_postcondition + reach_targets_precondition
+            states.append(
+                _get_next_state_spec(
+                    prev_state_spec=states[-1],
+                    new_state_id=new_state_id,
+                    transition=transitions[i],
+                    embodiment_id=embodiment_id,
+                    reach_targets=reach_targets,
+                    is_final_state=is_final_state,
+                )
+            )
+            out_tasks.append(
+                task.model_copy(
+                    update={"initial_state_spec_id": f"state_spec_{i}", "success_state_spec_id": new_state_id}
+                )
+            )
+
+        return ArenaEnvGraphSpec(
+            env_name=env_name or self.env_name,
+            nodes=self.nodes,
+            tasks=out_tasks,
+            state_specs=states,
+        )
 
     @property
     def nodes_by_id(self) -> dict[str, ArenaEnvGraphNodeSpec]:
@@ -191,3 +239,134 @@ class ArenaEnvGraphSpec(BaseModel):
         from isaaclab_arena.environments.arena_env_graph_conversion_utils import build_arena_env_from_graph_spec
 
         return build_arena_env_from_graph_spec(self)
+
+
+# --- Constraint resolution helpers (used by ArenaEnvGraphSpec.resolve_constraints) ---
+
+
+def _get_embodiment_id_from_nodes(nodes: list[ArenaEnvGraphNodeSpec]) -> str:
+    """Return the embodiment node's id (the parent of every reach constraint); the first one wins."""
+    ids = [node.id for node in nodes if node.type == ArenaEnvGraphNodeType.EMBODIMENT]
+    assert ids, "graph has no embodiment node"
+    return ids[0]
+
+
+def _get_task_state_transition(task: ArenaEnvGraphTaskSpec) -> TaskTransition:
+    """Look up the task class via ``TaskRegistry`` and return its declared success transition."""
+    task_cls = TaskRegistry().get_task_by_name(task.type)
+    assert task_cls is not None, f"task {task.type} not found in TaskRegistry"
+    return task_cls.success_state_transition(task.task_args)
+
+
+def _get_task_relocations(transition: TaskTransition) -> list[Relocate]:
+    """Return the transition's ``Relocate`` effects."""
+    for effect in transition.effects:
+        if not isinstance(effect, Relocate):
+            raise NotImplementedError(f"Effect {effect} is not yet supported.")
+    return list(transition.effects)
+
+
+def _get_next_state_spec(
+    prev_state_spec: ArenaEnvGraphStateSpec,
+    new_state_id: str,
+    transition: TaskTransition,
+    embodiment_id: str,
+    reach_targets: list[str | None],
+    is_final_state: bool,
+) -> ArenaEnvGraphStateSpec:
+    """Apply current task's success condition to the previous state spec to produce the next state spec.
+
+    Carries every constraint forward (re-prefixing its id to the new state id) except each
+    relocated object's old placement, which is replaced by what the success condition implies.
+    The final state spec additionally drops spawn-pose constraints (e.g. "at_position"), keeping only structural relations.
+    The drop is a semantic-cleanliness rule: spawn poses belong to reset-from states (they are
+    reset-time placement hints), structural relations belong to success states; the final state is
+    never reset-from, only checked as a success condition, so it is purely the latter.
+
+    Worked example -- task "place mug on bowl" (mug is the moved object), prev state holds::
+
+        mug at_position {x, y, z}          # mug's spawn pose
+        bowl on table                      # structural
+        bowl position_limits {x_min, ...}  # bowl's spawn pose (bowl is never moved)
+
+    yields, at an interior success state::
+
+        bowl on table                      # carried (unaffected)
+        bowl position_limits {...}         # carried (still a reset-from state)
+        mug on bowl                        # added (the relocation)
+        # mug at_position dropped -- replaced by the relocation above
+
+    and at the final state, additionally::
+
+        # bowl position_limits dropped -- final state keeps only structural relations
+
+    Args:
+        prev_state_spec: The previous state spec.
+        new_state_id: The id of the new state.
+        transition: The current task's declared success transition.
+        embodiment_id: The id of the embodiment.
+        reach_targets: The list of targets the embodiment must reach in this state.
+        is_final_state: Whether the state is the last one when reaching the final state.
+
+    Returns:
+        The next state spec.
+    """
+    relocations = _get_task_relocations(transition)
+    moved_objects_ids = {relocation.subject for relocation in relocations}
+    spatial_constraints: list[ArenaEnvGraphSpatialConstraintSpec] = []
+
+    for prev_spatial_constraint in prev_state_spec.spatial_constraints:
+        # Spawn pose constraints are those that are set at reset time. e.g."at position", "position_limits", etc.
+        constraint_is_spawn_pose = spatial_constraint_is_spawn_pose(prev_spatial_constraint.type)
+        # A moved object's old placement is replaced by the new placement
+        # Case 1: when the object is the constraint's child. (e.g. "cube on table" -> "cube on shelf")
+        # Case 2: when the object is the owner of its own (unary) spawn-pose constraint. (e.g. "cube at_position A" -> "cube at_position B")
+        child_is_relocated = prev_spatial_constraint.child in moved_objects_ids
+        parent_is_spawn_pose_owner = (prev_spatial_constraint.parent in moved_objects_ids) and constraint_is_spawn_pose
+        constraint_is_replaced = child_is_relocated or parent_is_spawn_pose_owner
+
+        # Because the spawn pose is at reset time, it is not affected by the success condition.
+        # e.g. a never-moved bowl's "position_limits".
+        constraint_is_dropped = is_final_state and constraint_is_spawn_pose
+
+        # Keep the old ones (spawn pose or structural) when they are not affected by the success condition.
+        if not (constraint_is_replaced or constraint_is_dropped):
+            spatial_constraints.append(
+                prev_spatial_constraint.model_copy(
+                    update={"id": _reprefix_id(prev_spatial_constraint.id, prev_state_spec.id, new_state_id)}
+                )
+            )
+    # Add the new ones.
+    for relocation in relocations:
+        spatial_constraints.append(
+            ArenaEnvGraphSpatialConstraintSpec(
+                id=f"{new_state_id}_{relocation.subject}_{relocation.relation}_{relocation.target}",
+                type=relocation.relation,
+                parent=relocation.target,
+                child=relocation.subject,
+            )
+        )
+    # One reach constraint per distinct target the embodiment must reach in this state.
+    # e.g. PnP A to B -> close door: B & door shall be reachable in this state.
+    seen_targets: set[str] = set()
+    task_constraints: list[ArenaEnvGraphTaskConstraintSpec] = []
+    for reach_target in reach_targets:
+        if reach_target is not None and reach_target not in seen_targets:
+            seen_targets.add(reach_target)
+            task_constraints.append(
+                ArenaEnvGraphTaskConstraintSpec(
+                    id=f"{new_state_id}_{embodiment_id.split('_')[0]}_reach_{reach_target}",
+                    type=ArenaEnvGraphTaskConstraintType.REACH,
+                    parent=embodiment_id,
+                    child=reach_target,
+                )
+            )
+    return ArenaEnvGraphStateSpec(
+        id=new_state_id, spatial_constraints=spatial_constraints, task_constraints=task_constraints
+    )
+
+
+def _reprefix_id(old_id: str, old_prefix: str, new_prefix: str) -> str:
+    """Swap a constraint id's ``{state_id}`` prefix so carried constraints stay uniquely named."""
+    assert old_id.startswith(old_prefix), f"constraint id {old_id!r} is not prefixed by its state id {old_prefix!r}"
+    return new_prefix + old_id[len(old_prefix) :]
